+++ apps/heft/src/plugins/TypeScriptPlugin/TypeScriptBuilder.ts
@@ -4,13 +4,9 @@
 import * as path from 'path';
 import * as semver from 'semver';
 import {
-  Colors,
-  Path,
   FileSystemStats,
   IFileSystemCreateLinkOptions,
-  IColorableSequence,
   Terminal,
-  ITerminalProvider,
   JsonFile,
   IPackageJson,
   InternalError
@@ -19,9 +15,6 @@ import * as crypto from 'crypto';
 import { Typescript as TTypescript } from '@microsoft/rush-stack-compiler-3.7';
 import {
   ExtendedTypeScript,
-  IEmitResolver,
-  IEmitHost,
-  IEmitTransformers,
   IExtendedProgram,
   IExtendedSourceFile,
   IResolveModuleNameResolutionHost
@@ -30,10 +23,12 @@ import {
 import { SubprocessRunnerBase } from '../../utilities/subprocess/SubprocessRunnerBase';
 import { Async } from '../../utilities/Async';
 import { PerformanceMeasurer, PerformanceMeasurerAsync } from '../../utilities/Performance';
-import { PrefixProxyTerminalProvider } from '../../utilities/PrefixProxyTerminalProvider';
 import { Tslint } from './Tslint';
 import { Eslint } from './Eslint';
-import { ISharedTypeScriptConfiguration, IEmitModuleKindBase } from '../../stages/BuildStage';
+import { ISharedTypeScriptConfiguration } from '../../stages/BuildStage';
+import { IScopedLogger } from '../../pluginFramework/logging/ScopedLogger';
+
+import { EmitFilesPatch, ICachedEmitModuleKind } from './EmitFilesPatch';
 
 export interface ITypeScriptBuilderConfiguration extends ISharedTypeScriptConfiguration {
   buildFolder: string;
@@ -45,7 +40,7 @@ export interface ITypeScriptBuilderConfiguration extends ISharedTypeScriptConfig
    * If provided, this is included in the logging prefix. For example, if this
    * is set to "other-tsconfig", logging lines will start with [typescript (other-tsconfig)].
    */
-  terminalPrefixLabel: string | undefined;
+  loggerPrefixLabel: string | undefined;
 
   lintingEnabled: boolean;
 
@@ -66,21 +61,6 @@ export interface ITypeScriptBuilderConfiguration extends ISharedTypeScriptConfig
    * The default is 50.
    */
   maxWriteParallelism: number;
-}
-
-interface ICachedEmitModuleKind<TModuleKind> extends IEmitModuleKindBase<TModuleKind> {
-  /**
-   * TypeScript's output is placed in the \<project root\>/.heft/build-cache folder.
-   * This is the the path to the subfolder in the build-cache folder that this emit kind
-   * written to.
-   */
-  cacheOutFolderPath: string;
-
-  /**
-   * Set to true if this is the emit kind that is specified in the tsconfig.json.
-   * Sourcemaps and declarations are only emitted for the primary module kind.
-   */
-  isPrimary: boolean;
 }
 
 type TWatchCompilerHost = TTypescript.WatchCompilerHostOfFilesAndCompilerOptions<
@@ -112,12 +92,14 @@ export class TypeScriptBuilder extends SubprocessRunnerBase<ITypeScriptBuilderCo
   private _typescriptParsedVersion: semver.SemVer;
 
   private _capabilities: ICompilerCapabilities;
+  private _useIncrementalProgram: boolean;
 
   private _eslintEnabled: boolean;
   private _tslintEnabled: boolean;
   private _moduleKindsToEmit: ICachedEmitModuleKind<TTypescript.ModuleKind>[];
   private _eslintConfigFilePath: string;
   private _tslintConfigFilePath: string;
+  private _typescriptLogger: IScopedLogger;
   private _typescriptTerminal: Terminal;
 
   private __tsCacheFilePath: string;
@@ -146,23 +128,12 @@ export class TypeScriptBuilder extends SubprocessRunnerBase<ITypeScriptBuilderCo
     return this.__tsCacheFilePath;
   }
 
-  public static getTypeScriptTerminal(
-    terminalProvider: ITerminalProvider,
-    prefixLabel: string | undefined
-  ): Terminal {
-    const prefixTerminalProvider: PrefixProxyTerminalProvider = new PrefixProxyTerminalProvider(
-      terminalProvider,
-      prefixLabel ? `[typescript (${prefixLabel})] ` : '[typescript] '
+  public async invokeAsync(): Promise<void> {
+    const loggerPrefixLabel: string | undefined = this._configuration.loggerPrefixLabel;
+    this._typescriptLogger = await this.requestScopedLoggerAsync(
+      loggerPrefixLabel ? `typescript (${loggerPrefixLabel})` : 'typescript'
     );
-
-    return new Terminal(prefixTerminalProvider);
-  }
-
-  public initialize(): void {
-    this._typescriptTerminal = TypeScriptBuilder.getTypeScriptTerminal(
-      this._terminalProvider,
-      this._configuration.terminalPrefixLabel
-    );
+    this._typescriptTerminal = this._typescriptLogger.terminal;
 
     // Determine the compiler version
     const compilerPackageJsonFilename: string = path.join(
@@ -192,6 +163,11 @@ export class TypeScriptBuilder extends SubprocessRunnerBase<ITypeScriptBuilderCo
       this._capabilities.incrementalProgram = true;
     }
 
+    // Disable incremental "useIncrementalProgram" in watch mode because its compiler configuration is
+    // different, which will invalidate the incremental build cache.  In order to support this, we'd need
+    // to delete the cache when switching modes, or else maintain two separate cache folders.
+    this._useIncrementalProgram = this._capabilities.incrementalProgram && !this._configuration.watchMode;
+
     this._configuration.buildCacheFolder = this._configuration.buildCacheFolder.replace(/\\/g, '/');
     this._tslintConfigFilePath = path.resolve(this._configuration.buildFolder, 'tslint.json');
     this._eslintConfigFilePath = path.resolve(this._configuration.buildFolder, '.eslintrc.js');
@@ -205,9 +181,7 @@ export class TypeScriptBuilder extends SubprocessRunnerBase<ITypeScriptBuilderCo
     if (this._eslintEnabled) {
       this._eslintEnabled = this._fileSystem.exists(this._eslintConfigFilePath);
     }
-  }
 
-  public async invokeAsync(): Promise<void> {
     // Report a warning if the TypeScript version is too old/new.  The current oldest supported version is
     // TypeScript 2.9. Prior to that the "ts.getConfigFileParsingDiagnostics()" API is missing; more fixups
     // would be required to deal with that.  We won't do that work unless someone requests it.
@@ -272,11 +246,14 @@ export class TypeScriptBuilder extends SubprocessRunnerBase<ITypeScriptBuilderCo
         throw new Error('Unable to resolve "tslint" package');
       }
 
+      const tslintScopedLogger: IScopedLogger = await this.requestScopedLoggerAsync(
+        loggerPrefixLabel ? `tslint (${loggerPrefixLabel})` : 'tslint'
+      );
       tslint = new Tslint({
         ts: ts,
         tslintPackagePath: this._configuration.tslintToolPath,
-        terminalPrefixLabel: this._configuration.terminalPrefixLabel,
-        terminalProvider: this._terminalProvider,
+        terminalPrefixLabel: this._configuration.loggerPrefixLabel,
+        scopedLogger: tslintScopedLogger,
         buildFolderPath: this._configuration.buildFolder,
         buildCacheFolderPath: this._configuration.buildCacheFolder,
         linterConfigFilePath: this._tslintConfigFilePath,
@@ -284,17 +261,21 @@ export class TypeScriptBuilder extends SubprocessRunnerBase<ITypeScriptBuilderCo
         measurePerformance: measureTsPerformance
       });
     }
+
     let eslint: Eslint | undefined = undefined;
     if (this._eslintEnabled) {
       if (!this._configuration.eslintToolPath) {
         throw new Error('Unable to resolve "eslint" package');
       }
 
+      const eslintScopedLogger: IScopedLogger = await this.requestScopedLoggerAsync(
+        loggerPrefixLabel ? `eslint (${loggerPrefixLabel})` : 'eslint'
+      );
       eslint = new Eslint({
         ts: ts,
         eslintPackagePath: this._configuration.eslintToolPath,
-        terminalPrefixLabel: this._configuration.terminalPrefixLabel,
-        terminalProvider: this._terminalProvider,
+        terminalPrefixLabel: this._configuration.loggerPrefixLabel,
+        scopedLogger: eslintScopedLogger,
         buildFolderPath: this._configuration.buildFolder,
         buildCacheFolderPath: this._configuration.buildCacheFolder,
         linterConfigFilePath: this._eslintConfigFilePath,
@@ -337,6 +318,8 @@ export class TypeScriptBuilder extends SubprocessRunnerBase<ITypeScriptBuilderCo
 
     this._validateTsconfig(ts, tsconfig);
 
+    EmitFilesPatch.install(ts, tsconfig, this._moduleKindsToEmit, /* useBuildCache */ false);
+
     ts.createWatchProgram(compilerHost);
 
     return new Promise(() => {
@@ -377,7 +360,7 @@ export class TypeScriptBuilder extends SubprocessRunnerBase<ITypeScriptBuilderCo
     let builderProgram: TTypescript.BuilderProgram | undefined = undefined;
     let tsProgram: TTypescript.Program;
 
-    if (this._capabilities.incrementalProgram) {
+    if (this._useIncrementalProgram) {
       builderProgram = ts.createIncrementalProgram({
         rootNames: tsconfig.fileNames,
         options: tsconfig.options,
@@ -411,8 +394,8 @@ export class TypeScriptBuilder extends SubprocessRunnerBase<ITypeScriptBuilderCo
     );
     //#endregion
 
-    //#region DIAGNOSTICS
-    const { duration: diagnosticsDurationMs, diagnostics } = measureTsPerformance('Diagnostics', () => {
+    //#region ANALYSIS
+    const { duration: diagnosticsDurationMs, diagnostics } = measureTsPerformance('Analyze', () => {
       const rawDiagnostics: TTypescript.Diagnostic[] = [
         ...genericProgram.getConfigFileParsingDiagnostics(),
         ...genericProgram.getOptionsDiagnostics(),
@@ -425,7 +408,7 @@ export class TypeScriptBuilder extends SubprocessRunnerBase<ITypeScriptBuilderCo
       );
       return { diagnostics: _diagnostics };
     });
-    this._typescriptTerminal.writeVerboseLine(`Diagnostics: ${diagnosticsDurationMs}ms`);
+    this._typescriptTerminal.writeVerboseLine(`Analyze: ${diagnosticsDurationMs}ms`);
     //#endregion
 
     //#region EMIT
@@ -598,16 +581,13 @@ export class TypeScriptBuilder extends SubprocessRunnerBase<ITypeScriptBuilderCo
     );
     //#endregion
 
-    let compilationFailed: boolean = false;
     if (diagnostics.length > 0) {
       this._typescriptTerminal.writeErrorLine(
-        `Encountered ${diagnostics.length} TypeScript error${diagnostics.length > 1 ? 's' : ''}:`
+        `Encountered ${diagnostics.length} TypeScript issue${diagnostics.length > 1 ? 's' : ''}:`
       );
       for (const diagnostic of diagnostics) {
-        this._printDiagnosticMessage(ts, diagnostic, true);
+        this._printDiagnosticMessage(ts, diagnostic);
       }
-
-      compilationFailed = true;
     }
 
     if (eslint) {
@@ -617,18 +597,10 @@ export class TypeScriptBuilder extends SubprocessRunnerBase<ITypeScriptBuilderCo
     if (tslint) {
       tslint.reportFailures();
     }
-
-    if (compilationFailed) {
-      throw new Error('TypeScript compilation failed.');
-    }
   }
 
-  private _printDiagnosticMessage(
-    ts: ExtendedTypeScript,
-    diagnostic: TTypescript.Diagnostic,
-    withIndent: boolean = false
-  ): void {
-    let terminalMessage: (string | IColorableSequence)[];
+  private _printDiagnosticMessage(ts: ExtendedTypeScript, diagnostic: TTypescript.Diagnostic): void {
+    let diagnosticMessage: string;
     // Code taken from reference example
     if (diagnostic.file) {
       const { line, character } = diagnostic.file.getLineAndCharacterOfPosition(diagnostic.start!);
@@ -637,35 +609,48 @@ export class TypeScriptBuilder extends SubprocessRunnerBase<ITypeScriptBuilderCo
         this._configuration.buildFolder,
         diagnostic.file.fileName
       );
-      terminalMessage = [
-        Colors.red(`ERROR: ${buildFolderRelativeFilename}:${line + 1}:${character + 1}`),
-        ` - (TS${diagnostic.code}) `,
-        Colors.red(message)
-      ];
+      diagnosticMessage =
+        `${buildFolderRelativeFilename}:${line + 1}:${character + 1} - ` +
+        `(TS${diagnostic.code}) ${message}`;
     } else {
-      terminalMessage = [ts.flattenDiagnosticMessageText(diagnostic.messageText, '\n')];
+      diagnosticMessage = ts.flattenDiagnosticMessageText(diagnostic.messageText, '\n');
     }
 
-    if (withIndent) {
-      terminalMessage = ['  ', ...terminalMessage];
-    }
+    const errorObject: Error = new Error(diagnosticMessage);
 
-    switch (diagnostic.category) {
+    const adjustedCategory: TTypescript.DiagnosticCategory = this._getAdjustedDiagnosticCategory(diagnostic);
+
+    switch (adjustedCategory) {
       case ts.DiagnosticCategory.Error: {
-        this._typescriptTerminal.writeErrorLine(...terminalMessage);
+        this._typescriptLogger.emitError(errorObject);
         break;
       }
 
       case ts.DiagnosticCategory.Warning: {
-        this._typescriptTerminal.writeWarningLine(...terminalMessage);
+        this._typescriptLogger.emitWarning(errorObject);
         break;
       }
 
       default: {
-        this._typescriptTerminal.writeLine(...terminalMessage);
+        this._typescriptTerminal.writeLine(...diagnosticMessage);
         break;
       }
     }
+  }
+
+  private _getAdjustedDiagnosticCategory(diagnostic: TTypescript.Diagnostic): TTypescript.DiagnosticCategory {
+    // Workaround for https://github.com/microsoft/TypeScript/issues/40058
+    // The compiler reports a hard error for issues such as this:
+    //
+    //    error TS6133: 'x' is declared but its value is never read.
+    //
+    // These should properly be treated as warnings, because they are purely cosmetic issues.
+    // TODO: Maybe heft should provide a config file for managing DiagnosticCategory mappings.
+    if (diagnostic.reportsUnnecessary && diagnostic.category === TTypescript.DiagnosticCategory.Error) {
+      return TTypescript.DiagnosticCategory.Warning;
+    }
+
+    return diagnostic.category;
   }
 
   private _emit(
@@ -675,119 +660,12 @@ export class TypeScriptBuilder extends SubprocessRunnerBase<ITypeScriptBuilderCo
   ): IExtendedEmitResult {
     const filesToWrite: IFileToWrite[] = [];
 
-    let foundPrimary: boolean = false;
-    let defaultModuleKind: TTypescript.ModuleKind;
-
-    for (const moduleKindToEmit of this._moduleKindsToEmit) {
-      if (moduleKindToEmit.isPrimary) {
-        if (foundPrimary) {
-          throw new Error('Multiple primary module emit kinds encountered.');
-        } else {
-          foundPrimary = true;
-        }
-        defaultModuleKind = moduleKindToEmit.moduleKind;
-      }
-    }
-
-    const baseEmitFiles: typeof ts.emitFiles = ts.emitFiles;
-
     const changedFiles: Set<IExtendedSourceFile> = new Set<IExtendedSourceFile>();
+    EmitFilesPatch.install(ts, tsconfig, this._moduleKindsToEmit, /* useBuildCache */ true, changedFiles);
 
-    let originalOutDir: string | undefined = undefined;
-    let redirectedOutDir: string | undefined = undefined;
     const writeFileCallback: TTypescript.WriteFileCallback = (filePath: string, data: string) => {
-      // Redirect from "path/to/lib" --> "path/to/.heft/build-cache/lib"
-      let redirectedFilePath: string = filePath;
-      if (redirectedOutDir !== undefined) {
-        if (Path.isUnderOrEqual(filePath, originalOutDir!)) {
-          redirectedFilePath = path.resolve(redirectedOutDir, path.relative(originalOutDir!, filePath));
-        } else {
-          // The compiler is writing some other output, for example:
-          // ./.heft/build-cache/ts_a7cd263b9f06b2440c0f2b2264746621c192f2e2.json
-        }
-      }
-
+      const redirectedFilePath: string = EmitFilesPatch.getRedirectedFilePath(filePath);
       filesToWrite.push({ filePath: redirectedFilePath, data });
-    };
-
-    // Override the underlying file emitter to run itself once for each flavor
-    // This is a rather inelegant way to convince the TypeScript compiler not to duplicate parse/link/check
-    ts.emitFiles = (
-      resolver: IEmitResolver,
-      host: IEmitHost,
-      targetSourceFile: IExtendedSourceFile | undefined,
-      emitTransformers: IEmitTransformers,
-      emitOnlyDtsFiles?: boolean,
-      onlyBuildInfo?: boolean,
-      forceDtsEmit?: boolean
-    ): TTypescript.EmitResult => {
-      if (onlyBuildInfo || emitOnlyDtsFiles) {
-        // There should only be one tsBuildInfo and one set of declaration files
-        return baseEmitFiles(
-          resolver,
-          host,
-          targetSourceFile,
-          emitTransformers,
-          emitOnlyDtsFiles,
-          onlyBuildInfo,
-          forceDtsEmit
-        );
-      } else {
-        if (targetSourceFile) {
-          changedFiles.add(targetSourceFile);
-        }
-
-        let defaultModuleKindResult: TTypescript.EmitResult;
-        let emitSkipped: boolean = false;
-        for (const moduleKindToEmit of this._moduleKindsToEmit) {
-          const compilerOptions: TTypescript.CompilerOptions = moduleKindToEmit.isPrimary
-            ? {
-                ...tsconfig.options
-              }
-            : {
-                ...tsconfig.options,
-                module: moduleKindToEmit.moduleKind,
-
-                // Don't emit declarations for secondary module kinds
-                declaration: false,
-                declarationMap: false
-              };
-
-          if (!compilerOptions.outDir) {
-            throw new InternalError('Expected compilerOptions.outDir to be assigned');
-          }
-
-          // Redirect from "path/to/lib" --> "path/to/.heft/build-cache/lib"
-          originalOutDir = compilerOptions.outDir;
-          redirectedOutDir = moduleKindToEmit.cacheOutFolderPath;
-
-          const flavorResult: TTypescript.EmitResult = baseEmitFiles(
-            resolver,
-            {
-              ...host,
-              getCompilerOptions: () => compilerOptions
-            },
-            targetSourceFile,
-            ts.getTransformers(compilerOptions, undefined, emitOnlyDtsFiles),
-            emitOnlyDtsFiles,
-            onlyBuildInfo,
-            forceDtsEmit
-          );
-
-          emitSkipped = emitSkipped || flavorResult.emitSkipped;
-          if (moduleKindToEmit.moduleKind === defaultModuleKind) {
-            defaultModuleKindResult = flavorResult;
-          }
-
-          originalOutDir = undefined;
-          redirectedOutDir = undefined;
-          // Should results be aggregated, in case for whatever reason the diagnostics are not the same?
-        }
-        return {
-          ...defaultModuleKindResult!,
-          emitSkipped
-        };
-      }
     };
 
     const result: TTypescript.EmitResult = genericProgram.emit(
@@ -795,7 +673,7 @@ export class TypeScriptBuilder extends SubprocessRunnerBase<ITypeScriptBuilderCo
       writeFileCallback
     );
 
-    ts.emitFiles = baseEmitFiles;
+    EmitFilesPatch.uninstall(ts);
 
     return {
       ...result,
@@ -930,7 +808,7 @@ export class TypeScriptBuilder extends SubprocessRunnerBase<ITypeScriptBuilderCo
       currentFolder
     );
 
-    if (this._capabilities.incrementalProgram) {
+    if (this._useIncrementalProgram) {
       tsconfig.options.incremental = true;
       tsconfig.options.tsBuildInfoFile = this._tsCacheFilePath;
     }
@@ -944,7 +822,7 @@ export class TypeScriptBuilder extends SubprocessRunnerBase<ITypeScriptBuilderCo
   ): TTypescript.CompilerHost {
     let compilerHost: TTypescript.CompilerHost;
 
-    if (this._capabilities.incrementalProgram) {
+    if (this._useIncrementalProgram) {
       compilerHost = ts.createIncrementalCompilerHost(tsconfig.options);
     } else {
       compilerHost = ts.createCompilerHost(tsconfig.options);
@@ -992,16 +870,25 @@ export class TypeScriptBuilder extends SubprocessRunnerBase<ITypeScriptBuilderCo
       (
         rootNames: ReadonlyArray<string> | undefined,
         options: TTypescript.CompilerOptions | undefined,
-        host?: TTypescript.CompilerHost,
+        compilerHost?: TTypescript.CompilerHost,
         oldProgram?: TTypescript.EmitAndSemanticDiagnosticsBuilderProgram,
         configFileParsingDiagnostics?: ReadonlyArray<TTypescript.Diagnostic>,
         projectReferences?: ReadonlyArray<TTypescript.ProjectReference> | undefined
       ) => {
-        // TODO: Support additionalModuleKindsToEmit
+        if (compilerHost === undefined) {
+          throw new InternalError('_buildWatchCompilerHost() expects a compilerHost to be configured');
+        }
+
+        const originalWriteFile: TTypescript.WriteFileCallback = compilerHost.writeFile;
+        compilerHost.writeFile = (filePath: string, ...rest: unknown[]) => {
+          const redirectedFilePath: string = EmitFilesPatch.getRedirectedFilePath(filePath);
+          originalWriteFile.call(this, redirectedFilePath, ...rest);
+        };
+
         return ts.createEmitAndSemanticDiagnosticsBuilderProgram(
           rootNames,
           options,
-          host,
+          compilerHost,
           oldProgram,
           configFileParsingDiagnostics,
           projectReferences