+++ apps/heft/src/utilities/subprocess/SubprocessTerminalProvider.ts
@@ -4,17 +4,22 @@
 import { ITerminalProvider, TerminalProviderSeverity } from '@rushstack/node-core-library';
 
 import { ISubprocessInnerConfiguration } from './SubprocessRunnerBase';
-import { ISubprocessLoggingMessage } from './SubprocessCommunication';
 
 export class SubprocessTerminalProvider implements ITerminalProvider {
   private _builderConfiguration: ISubprocessInnerConfiguration;
 
-  public constructor(builderConfiguration: ISubprocessInnerConfiguration) {
+  public write: (data: string, severity: TerminalProviderSeverity) => void;
+
+  public constructor(
+    builderConfiguration: ISubprocessInnerConfiguration,
+    writeFn: (data: string, severity: TerminalProviderSeverity) => void
+  ) {
     if (!process.send) {
       throw new Error(`The process.send function is not supported in this context`);
     }
 
     this._builderConfiguration = builderConfiguration;
+    this.write = writeFn;
   }
 
   public get supportsColor(): boolean {
@@ -23,14 +28,5 @@ export class SubprocessTerminalProvider implements ITerminalProvider {
 
   public get eolCharacter(): string {
     return this._builderConfiguration.terminalEolCharacter;
-  }
-
-  public write(data: string, severity: TerminalProviderSeverity): void {
-    const message: ISubprocessLoggingMessage = {
-      type: 'logging',
-      data,
-      severity
-    };
-    process.send!(message);
   }
 }