+++ build-tests/heft-webpack-basic-test/webpack.config.js
@@ -32,6 +32,12 @@ function createWebpackConfig({ production }) {
       path: path.join(__dirname, 'dist'),
       filename: '[name]_[contenthash].js'
     },
+    performance: {
+      // This specifies the bundle size limit that will trigger Webpack's warning saying:
+      // "The following entrypoint(s) combined asset size exceeds the recommended limit."
+      maxEntrypointSize: 250000,
+      maxAssetSize: 250000
+    },
     devtool: production ? undefined : 'source-map',
     plugins: [
       // See here for documentation: https://github.com/jantimon/html-webpack-plugin