+++ libraries/debug-certificate-manager/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log - @rushstack/debug-certificate-manager
 
-This log was last generated on Thu, 13 Aug 2020 04:57:38 GMT and should not be manually modified.
+This log was last generated on Thu, 13 Aug 2020 09:26:40 GMT and should not be manually modified.
+
+## 0.2.33
+Thu, 13 Aug 2020 09:26:40 GMT
+
+*Version update only*
 
 ## 0.2.32
 Thu, 13 Aug 2020 04:57:38 GMT