+++ libraries/package-deps-hash/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log - @rushstack/package-deps-hash
 
-This log was last generated on Thu, 13 Aug 2020 04:57:38 GMT and should not be manually modified.
+This log was last generated on Thu, 13 Aug 2020 09:26:40 GMT and should not be manually modified.
+
+## 2.4.37
+Thu, 13 Aug 2020 09:26:40 GMT
+
+*Version update only*
 
 ## 2.4.36
 Thu, 13 Aug 2020 04:57:38 GMT