+++ webpack/loader-raw-script/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log - @rushstack/loader-raw-script
 
-This log was last generated on Thu, 13 Aug 2020 04:57:38 GMT and should not be manually modified.
+This log was last generated on Thu, 13 Aug 2020 09:26:40 GMT and should not be manually modified.
+
+## 1.3.36
+Thu, 13 Aug 2020 09:26:40 GMT
+
+*Version update only*
 
 ## 1.3.35
 Thu, 13 Aug 2020 04:57:38 GMT